refactor(asset): keep abi/mod.rs as pure glue, move logic into v2.rs
Addresses review feedback on #4218: logic does not belong in mod.rs. Move the
`as_label` impl, the ERC-165 interface-id consts, and the cross-surface test
module out of `b20_asset/abi/mod.rs` and into `abi/v2.rs` (the canonical
surface file, which owns anything keyed to the newest surface). `abi/mod.rs`
is now only module docs plus `mod`/`pub use` re-exports.

Pure code movement: fingerprint layout and values are unchanged, all tests
pass without re-blessing.
